Pull out clearing innerHTML
