+++ gui/src/app/Scripting/ScriptEditor.tsx
@@ -184,22 +184,4 @@ const ConsoleOutputWindow: FunctionComponent<ConsoleOutputWindowProps> = ({
   return <div className="ConsoleOutputArea" ref={consoleRef} />;
 };
 
-type ConsoleOutType = "stdout" | "stderr";
-
-export const writeConsoleOutToDiv = (
-  parentDiv: RefObject<HTMLDivElement>,
-  x: string,
-  type: ConsoleOutType,
-) => {
-  if (x === "") return;
-  if (!parentDiv.current) return;
-  const styleClass = type === "stdout" ? "WorkerStdout" : "WorkerStderr";
-  const preElement = document.createElement("pre");
-  preElement.textContent = x;
-  const divElement = document.createElement("div");
-  divElement.className = styleClass;
-  divElement.appendChild(preElement);
-  parentDiv.current.appendChild(divElement);
-};
-
 export default ScriptEditor;
